requirements.txt
cffi>=1.17.1
cryptography>=44.0.0
psycopg>=3.2.3
pycparser>=2.22
pycryptodome>=3.21.0
typing_extensions>=4.12.2
